backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
opencv-python>=4.8.0
motor>=3.3.1
certifi>=2023.7.22
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-jose[cryptography]>=3.3.0
